HTTPRequest Maximum Header/Body Length
Added fields to the HTTPRequest class, and functionality to it internally, to support specifying a maximum size -- in bytes -- of both the header and body of the response.  If the remote server attempts to send more than the specified maximum, an exception is raised and the request is aborted at once.