JSON Multi-Byte String Fix
In checking to see if input to the JSON::Decode function is simply the string "null" -- to avoid ambiguity in json_decode's return value -- it makes sense to use strtolower since the string "null" contains no non-ASCII characters.  However, it does not make sense to use the ASCII version of the trim function, since the input could very well be "null" surrounded by Unicode whitespace characters.  This has been corrected.