JSON Client Maximum Body/Header Size Support
Added support for the newly-added maximum_header_bytes and maximum_body_bytes fields of the HTTPRequest class to the JSONClient class.  The JSONClient uses a default value of 65536 (i.e. 64 kilobytes) for these values, and the user may adjust them by accessing the JSONClient's now-public request field (which contains the HTTPRequest object the JSONClient will use to send requests).