Base64
Added a Base64 class, which contains the static function Encoding and Decode, which wrap the PHP functions base64_encode and base64_decode, respectively.

In the case of Base64::Decode the wrapper forces strict behaviour, unlike base64_decode, which is not strict by default.  There's also a check to ensure the input base64 string does not contain whitespace -- which is invalid and in violation of the base64 spec (base64_decode happily accepts it, however, even in "strict" mode).