URL Handling Utilities
Added some utilities for dealing with URLs under the URL class:

The Escape function which -- in keeping with the theme of the HTML::Escape function -- takes an arbitrary number of arguments, escapes and concatenates all of them.

The GetQueryString function which converts an associative array into a query string.

The Combine function which combines two URL components.

Refactored several places in the code base to use functions from the URL class rather than rolling their own/using PHP built-in functions.