CanadaPost\Dimensions Input Sanification
The Canada Post XML API has a very particular idea of what "length", "width", and "height" mean.  I.e. the length, width, and height -- as far as the API is concerned -- are the longest, second longest, and shortest dimensions of an object, respectively.

To respect this ordering, modified the CanadaPost\Dimensions constructor to automatically reorder the $length, $width, and $height parameters if they are not ordered properly.