String Fix
The PHP multi-byte string extension does not default to using UTF-8 as its internal encoding.

Accordingly I modified the string.php file so that whenever it is included -- as it must be before any of the multi-byte string helpers are used -- it sets the multi-byte string extension's internal encoding to UTF-8.

As with the rest of the multi-byte string utilities, this shim functions even if the PHP multi-byte string extension is not installed -- it checks to make sure the function mb_internal_encoding exists before calling it to change the encoding.