String Multi-Byte Check
Added the IsMultiByte function to the String class, which allows a caller to check to see if the String class is fully capable of handling multi-byte strings.  The function simply checks for the existence of all the functions and classes that the class uses to handle multi-byte strings, returns true if they're all found, false otherwise.