Test Front-End
Wrote a test driver and templates to display/format the result of tests.

The test driver looks at every .php file in the "tests" directory and tries to instantiate a class in the Tests namespace with the name of the file (minus the extension) which extends \CivicInfoBC\Testing\TestSuite.  If such a class is found, it calls the Execute method and stores the results, which are then formatted by the test.phtml template (which is rendered once per result set).