Crypto PRNG Fix
Added a shim to the Crypto::Random method to simply return the empty string if zero bytes are requested.  Proceeding and attempting to acquire zero bytes causes an exception when using OpenSSL as it fails to correctly report that the random number generator is suitable for use in cryptographic applications.