URL Fix
The URL::MakeQueryString method didn't set the $retr variable to the empty string unconditionally, instead assuming that the provided array had at least one element, and that the foreach loop would therefore executing at least once, initializing and setting $retr to "?".

Obviously this was a poor assumption, and it has been fixed.