Router Simplification
Simplified the code that creates a controller instance in the Router class.  Previously call_user_func_array was used to pass an array of arguments through to the ReflectionClass::newInstance method, however the ReflectionClass class has a method specifically for this use case: ReflectionClass::newInstanceArgs.

Accordingly that method is now used instead.