Base64 Unit Tests
Added 3 unit tests for the \CivicInfoBC\Base64 class.  The class-in-question is a very thin wrapper around the PHP base64_encode and _decode methods, so minimal tests are necessary, but a guard was added to reject strings containing whitespace (which PHP does not do).  Insuring this works is one of the unit tests.