HTTP Response Body Encoding
Added a method -- GetBody -- to the HTTP\Response class which searches the response headers for an appropriate header (a "Content-Type" header with "charset" set) to determine the encoding of the response's body.  It then uses the String\ConvertFrom method to transcode the body to UTF-8 and return it.  If an appropriate header is not found, UTF-8 is used as a fallback.

This allows for the seamless consumption of HTTP resources that do not use UTF-8, while simultaneously leaving the raw bytes of the body available.

Refactored the places where the HTTP\Response class is used to use the GetBody method rather than accessing the $body field directly.