Unit Tests Fail on PHP Error/Warning
Modified the \CivicInfoBC\Testing\Test::Execute method so that PHP errors and warnings are captured and cause the test to fail.

This means that even if the unit tests are running on a machine/in an environment where errors and warnings aren't passed through to the user, the unit tests will still fail.