HTTP\ContentType
Added a class to parse HTTP Content-Type headers -- HTTP\ContentType.

The constructor is optionally passed a header to parse.  If none is provided it will automatically fetch the Content-Type header for the currently executing request.

A default media type and character set may also be provided to avoid spurious conditionals to coalesce nulls to a sensible default.

Updated the HTTP\Response class to use this to obtain the appropriate character set to decode the body of the response it encapsulates.