HTTP\Request Unit Tests
Added 7 unit tests for the HTTP\Request class.

Because this class makes external HTTP requests, added a plain text file and a very simple script to which the class can make requests for testing purposes:

The request_test.txt file contains a fixed string, which can be compared to the body of the HTTP\Response object the HTTP\Request instance returns.

The request_test.php file contains a simple script which converts a fixed Unicode string to some encoding to test the HTTP\Request::GetBody method's character set detection and conversion.

In order to include request_test.php .gitignore had to be updated.