MySQL\Name Fix
All of the MySQL classes used to be under the \CivicInfoBC namespace, rather than the \CivicInfoBC\MySQL namespace.  Therefore they could use classes from the \CivicInfoBC namespace without qualification.

A reference to the Regex class in the \CivicInfoBC namespace in the MySQL\Name class escaped notice when the code was moved to the \CivicInfoBC\MySQL namespace, which triggered a PHP fatal error.

This has been fixed.