MySQL\Lock
Added the MySQL\Lock class, which provides an RAII wrapper around a pair of LOCK TABLES/UNLOCK TABLES SQL statements, acquiring the lock when the object is constructed, and releasing the lock when the last reference to the object goes out of scope.

Also added the MySQL\LockInfo class, which allows locks to be precisely specified (read vs. write).