Measure
Since the code for the Eli Mina form is getting into the realm of units, added the Measure class, which is a base class which encapsulates the idea of a measurement taken using a convertible quantity of units.

Derived classes supply a static conversion table which allows units to be converted (using late static binding).