ArrayUtil::Coalesce
Added the Coalesce static method to the ArrayUtil class, which takes some value, and returns it unchanged if it's an array, otherwise wraps it in an array and then returns it.

This method easily allows methods to take and operate on either a single value, or an array of such values -- the In static method did this internally previously, this has been switched to use the Coalesce method instead.