ArrayUtil::Find
Added the Find static method to the ArrayUtil class.  This method takes an array, a value, and optionally a boolean value and a comparison functor, and returns either the key at which the value exists in the array, or null if the value was not found.  The boolean value may be used to tell the function to throw rather than returning null.