JSONValidator
Implemented a generic JSON validation mechanism in the form of the JSONValidator class.

When creating an instance of the JSONValidator class it is supplied with a "schema" -- a string or some key/value mapping which identifies the object structure and types that should be present.

The Execute method then takes the result of JSON::Decode and ensures that it conforms to the schema, by either returning true/false, or returning a list of problems, depending on whether detailed output is enabled or not.