ArrayUtil::In Fix/Improvements
Improved/Fixed (depending on your understanding of "fixed") ArrayUtil::In's behaviour.

Previously ArrayUtil::In used the PHP built-in function in_array in the absence of a user-supplied comparison function.

However, when comparing strings, in_array -- even in strict mode -- will not find equivalent Unicode strings to be equal.

Accordingly, ArrayUtil::In is no longer implemented in terms of in_array -- user-supplied comparison function or not -- and uses String::Equals to compare strings.