Event
Implemented a base class for events which provides useful/convenient/shared functionality.

Also implemented a special exception type to be thrown when a registration fails for reasons unrelated to the database.  This allows such errors to be easily picked out and turned into an appropriate response.

Some documentation is missing.