Event::AddCommon
Added the AddCommon method to the Event class.

This method adds fields and default values common to all events to an object.

Since some events will have exceptional table structures, this is opt-in -- i.e. it is not called automatically from the Register method, the consumer must opt to call it to retrieve these common fields/values.