Event::CheckSession
Added a method to the Event class to retrieve the number of people enrolled in a particular session, based on a database column and a corresponding value in that column.

As this method exposes raw implementation details (underlying database columns et cetera) it is protected, meaning that derived classes have to wrap it if the functionality is to be visible to the outside.