MySQL\Connection Fix
Fixed a bug in the MySQL\Connection::Execute method which was causing queries with no parameters (i.e. whose query text did not require substitution) to fail.

The bug stemmed from the manner in which incoming arguments are forwarded through to sprintf.  An array of arguments is formed, with the first argument being $text.  Therefore, if there are no parameters, the count of this array is 1.

If there are no parameters, it is unnecessary to call sprintf, and accordingly a ternary conditional selects between passing through $text untouched, and formatting it with sprintf.  However this condition erroneously checked count($arr)===0 and was thus always false (since the method takes at least one argument).

The fix was to change this to count($arr)===1.