MySQL\Connection::Update Fixes
The MySQL\Connection::Update method was referring to the String class as simply "String", which was erroneous as the class-in-question is in the nested namespace \CivicInfoBC\MySQL, therefore qualification was needed.

This method also failed to add commas between statements in the SET clause of the UPDATE query it generated.

Lastly, the method applied count to incoming objects, even though they may not have been countable.  Changed this so that rather than calling count to obtain the number of values, and then building the query, the query is built, the values are counted, and if there are no values an exception is thrown.  The functionality/result is the same, but the implementation is more general -- all iterable objects may be handled, not just those that are countable.