Router Fixes
Fixed two bugs with the Router class:

During routing, if no controller or arguments were specified in the request, the Router would attempt to preserve the controller as an argument before attempting the default controller.  This would result in a spurious null entry in the arguments array.

During routing, the controller name was stored in the RouteInfo object prefixed with the search namespace.  This could lead to arguments being erroneously prefixed with a namespace if the controller was later converted back to an argument due to the fact no matching class was found.