Object Tweak
The Object class now persists null values that are explicitly assigned to it, rather than transforming them into unset operations.

This means that when one iterates on an Object, all the null values assigned will be retrieved, rather than the associated keys simply not being present.

This is important for instances where Object is used as an encapsulation for database rows.