Links Tweak
Links::Get now applies ArrayUtil::Coalesce to its $args and $get arguments, which means that if only one item is being supplied, the caller no longer needs to wrap that item in an array.

Also, the arguments are now defaulted to null rather than array(), which is a more natural way for the caller to "skip" an argument (i.e. not provide its value but provide the value of the one after it).