EventRegisterOptions and Allow Event Capacity to be Exceeded
Broke the extra, optional arguments to the Event::Register method out into their own class -- EventRegisterOptions.

This will allow these options to be extended easily, and in a backwards compatible manner.  It will also making specifying arguments simpler as the number of arguments grows (i.e. one will not have to manually pass all the defaults that come before the argument one wishes to customize).

In addition to $verify and $case_insensitive, which were already optional arguments to Event::Register, added $exceed_capacity, which suppresses the registration count check, allowing registrations to be performed above and beyond the event's capacity.