CrossOriginResourceSharing
Added the \CivicInfoBC\CrossOriginResourceSharing class, which allows for simple server side implementation of cross-origin resource sharing.

The class takes an array of allowed origin domains in its constructor, and when its Send method is invoked checks the "origin" header for a match.  If there is a match, it sends the "access-control-allow-origin" header, set to the appropriate value, thereby allowing the request.