ErrorHandler
Added the ErrorHandler class, which is a base and default/null implementation for objects which can handle errors.

The ErrorPage class -- which currently displays an error page and sends an e-mail -- will be broken out into two classes, both of which derive from this class, and which use the Decorator pattern to provide highly customizable, modular error handling.