HTML Tweaks
The HTML class can now be used statically (through the Escape and Write static members) or non-statically.  When used non-statically, it is constructed with some number of objects.  Whenever it is converted to a string those objects are escaped (as through a call to HTML::Escape) and the resulting string is returned.