Framework\Request Additions
Added the Framework\Request::Header method, which functions as the other Framework\Request getters but for HTTP header key/value pairs.

Framework\Request::Get, Framework\Request::Post, Framework\Request::Cookie, and Framework\Request::Header now allow their first parameter to be null (defaulted) and when this is the case return a Traversable which allows all key/value pairs of the collection-in-question to be traversed.